Refactor ShellV2 page into components and hooks
Splits the monolithic `ShellV2` component into smaller, manageable components:
- `ShellHeader`: Displays breadcrumbs and title.
- `ShellStatusBar`: Displays portal status and callback timer.
- `ShellTerminal`: Wraps xterm.js terminal and completions.
- `ShellCompletions`: Renders the autocomplete dropdown.

Extracts logic into custom hooks:
- `useShellData`: Manages GraphQL queries for shell, beacon, and portal data.
- `useCallbackTimer`: Calculates time until next callback.
- `useShellTerminal`: Encapsulates terminal initialization, WebSocket handling, and input state.

Updates `index.tsx` to compose these components.
Includes updated build artifacts as required by the repo structure.
